+++ packages/ui/voo_forms/lib/src/presentation/widgets/atoms/internal/async_dropdown_field_widget.dart
@@ -70,6 +70,7 @@ class AsyncDropdownFieldWidgetState<T> extends State<AsyncDropdownFieldWidget<T>
       sortComparator: widget.field.sortOptions,
       asyncSearch: widget.field.asyncOptionsLoader,
       searchDebounce: widget.field.searchDebounce,
+      optionBuilder: widget.field.optionBuilder,
       decoration: widget.field.getInputDecoration(context).copyWith(
         suffixIcon: widget.field.dropdownIcon,
       ),